requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.3
djangorestframework-simplejwt==5.3.1
redis==5.1.0
drf-yasg==1.21.7
django-cors-headers==4.4.0
coverage==7.6.1